feat(gabriel)

+++ Library/src/main/java/com/example/Library/service/UserService.java
@@ -23,23 +23,35 @@ public class UserService {
     }
     
     public User getUserById(Integer id){
+        
+        if (id == null || id == 0) throw new IllegalArgumentException(); 
+
         return userRepository.findById(id).orElse(null);
     }
 
     public void createUser(User user){
 
-        System.out.println(user.toString());
+        if (userVerification(user)) {
 
-        if (userRepository.existsByEmail(user.getEmail()) == true) throw new RuntimeException("O email já existe"); 
-
-        user.setStatus(true);
-        user.setLoans(new ArrayList<>());
-        user.setSingleCard(UUID.randomUUID());
-        userRepository.save(user);
+            user.setStatus(true);
+            user.setLoans(new ArrayList<>());
+            user.setSingleCard(UUID.randomUUID());
+            userRepository.save(user);
+        }
     }
 
     public void deleteUser(Integer id){
+
+        if (id == null || id == 0) throw new IllegalArgumentException(); 
+
         userRepository.deleteById(id);
     }
 
+    public Boolean userVerification(User user){
+
+        if(user.getName() == null || user.getName().isBlank()) throw new IllegalArgumentException();
+        if(user.getEmail() == null || user.getEmail().isBlank() || userRepository.existsByEmail(user.getEmail()) == true) throw new IllegalArgumentException();
+
+        return true;
+    }
 }
